Fix for rendering hooks warning
